Apres discussion avec Sylvain, la correction du bug de BlastX sur les types rentrent dans le repository

+++ eugene/src/SensorPlugins/BlastX/Sensor.BlastX.h
@@ -9,13 +9,13 @@
 class SensorBlastX : public Sensor
 {
  private:
-  float *ProtMatch, *ProtMatchLevel;
+  REAL *ProtMatch, *ProtMatchLevel;
   int  *ProtMatchPhase;
   std::vector<int>  vPos,     vPMPhase;
   std::vector<REAL> vPMLevel, vPMatch;
   std::vector<int>::iterator iter;
   int    index;
-  double keyBXLevel[10];
+  REAL keyBXLevel[10];
   int    minIn;
   int blastxM;
